fixed help pagination, renamed code to src
- fixed the help command react pagination controls.
- renamed the `code` folder to `src` because the pycharm debugger is stupid. https://stackoverflow.com/questions/29172399/unable-to-debug-in-pycharm-because-of-an-importerror-in-pydevconsole-py/42264566#42264566
- fixed a typo in the helptext documentation.
