+++ src/cogs/base.py
