+++ src/cogs/user_tools.py
@@ -2,7 +2,7 @@
 from enum import Enum
 from typing import Union
 
-from code.cogs.base import ConfiguredCog
+from src.cogs.base import ConfiguredCog
 
 from discord.ext import commands
 from discord import Role, Embed
